Move observer implementation to new derive crate.
Both Binance and Huobi observers had the same basic implementation,
the only difference were in the connection function, to prepare for
adding new observers this issue had to be solved.

This implementation uses derive macro to implement the currently the
most efficient version of websocket observer, it assumes the exact
fields in the observer structure that it's derived on, but both
observers already had the exactly same ABI anyway.

This also simplifies the implementation of observer to just implementing
connection function over WebSocket.